Completed Day 05
Today we created a really cool image gallery using mostly Flexbox and CSS transitions, and a little bit of JavasScript to toggle classes to elements on the page as a response to events. I was not very familiar with any of those things, but they were all pretty straight forward and useful, so this has been one of my favourites so far.